added

+++ Leetcode/Scala/0828. Count Unique Characters of All Substrings of a Given String/Solution/src/main/scala/Solution.scala
@@ -5,10 +5,10 @@
  object Solution {
     def uniqueLetterString(s: String): Int = {
         s.zipWithIndex.foldLeft((0, 0, Array.fill(128)(0), Array.fill(128)(0))) { case ((res, sub, prev, curr), (char, index)) =>
-            val newSub = sub + (index + 1) + prev(char.toInt) - 2 * curr(char.toInt)
+            val newSub = sub + (index + 1) + prev(char) - 2 * curr(char)
             val newRes = res + newSub
-            prev(char.toInt) = curr(char.toInt)
-            curr(char.toInt) = index + 1
+            prev(char) = curr(char)
+            curr(char) = index + 1
             (newRes, newSub, prev, curr)
         }.head
     }
